Handle properties of G_TYPE_INTERFACE type
If an object has a property which is of type G_TYPE_INTERFACE, we
just serialize the object implementing that interface. This is
the same as we do for inheritance, i.e. if the property specifies
that it is of type MyBase and it is currently holding a MyDerived,
then the serialized output will contain a MyDerived.

For both inheritance and interfaces, this requires that the
derived GType name is available at deserialization time, and of
course implies that the derived type can be deserialized properly.
If this is not desirable (for example if you are using a GInterface
for your public API but your private implementation is unstable),
then you'll need to customise the serialisation in some way.

At least, when it's possible to do custom serialisation, which is
not yet...