Add supported (certain) boxed types
Specifically, string arrays (GStrv) and GBytes, mostly because
these have 1-1 GVariant representations.

Well, almost: it turns out that GBytes is fine with embedded
NULs, and I think GVariant's (ay) should be too, but the
API to get a "bytestring" doesn't support it, so we work around
that.

Also, add a boxed type test